requirements.txt
alembic==1.4.2
astroid==2.4.2
bcrypt==3.2.0
blinker==1.4
cffi==1.14.3
click==7.1.2
colorama==0.4.3
coverage==5.3
cryptography==3.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
Flask-WTF==0.14.3
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
livereload==2.6.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
nose2==0.9.2
numpy==1.19.1
pandas==1.1.0
passlib==1.7.4
pycparser==2.20
pylint==2.5.3
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3